Fix test arrangement
